Isolate legacy code
