+++ Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-chrono = { version = "0.4.24", features = ["serde"]}
+chrono = { version = "0.4.24", features = ["serde"] }
 serde_json = "1.0.95"
 serde = { version = "1.0.159", features = ["derive"] }
 rand = "0.8.5"
