+++ src/cli/mod.rs
@@ -1,0 +1,5 @@
+pub use remaining_operation::remaining_operation;
+
+mod legacy_remaining_money_screen;
+mod remaining_operation;
+
