+++ src/main.rs
@@ -1,3 +1,5 @@
+use crate::cli::remaining_operation;
+
 mod accounts;
 mod amounts;
 mod cli;
@@ -7,7 +9,6 @@ mod remaining;
 mod transaction;
 mod vault;
 
-use crate::cli::remaining_operation;
 fn main() {
     remaining_operation()
 }
