+++ src/period/periods.rs
@@ -1,9 +1,10 @@
-use chrono::{Datelike, Months, NaiveDate};
-use serde::de::{Error as DeError, Unexpected, Visitor};
-use serde::ser::Error as SerError;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::fmt::Formatter;
 use std::str::FromStr;
+
+use chrono::{Datelike, Months, NaiveDate};
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use serde::de::{Error as DeError, Unexpected, Visitor};
+use serde::ser::Error as SerError;
 
 #[derive(Debug, PartialEq, Eq, Clone, Hash)]
 pub struct Period {
@@ -28,26 +29,32 @@ impl<'de> Deserialize<'de> for Period {
             where
                 E: DeError,
             {
-                fn split<'a, E: DeError>(s: &'a str, splitter: &str, exp: &str) -> Result<(&'a str, &'a str), E> {
+                fn split<'a, E: DeError>(
+                    s: &'a str,
+                    splitter: &str,
+                    exp: &str,
+                ) -> Result<(&'a str, &'a str), E> {
                     let split = s.split(splitter).collect::<Vec<&str>>();
                     if split.len() != 2 {
-                        return Err(DeError::invalid_value(
-                            Unexpected::Str(&s),
-                            &exp
-                        ))
+                        return Err(DeError::invalid_value(Unexpected::Str(&s), &exp));
                     }
                     Ok((split[0], split[1]))
                 }
                 let (start_date, end_date) = if v.len() == 21 {
-                    let (raw_start, raw_end) = split(&v, "/", "Two ISO dates separated by a slash",)?;
+                    let (raw_start, raw_end) =
+                        split(&v, "/", "Two ISO dates separated by a slash")?;
                     fn parse_date<E: DeError>(raw: &str) -> Result<NaiveDate, E> {
-                        NaiveDate::parse_from_str(raw, "%F").or(Err(DeError::invalid_value(Unexpected::Str(raw), &"an ISO date")))
+                        NaiveDate::parse_from_str(raw, "%F").or(Err(DeError::invalid_value(
+                            Unexpected::Str(raw),
+                            &"an ISO date",
+                        )))
                     }
                     let start_date = parse_date(raw_start)?;
                     let end_date = parse_date(raw_end)?;
                     (start_date, end_date)
                 } else if v.len() == 7 {
-                    let (raw_year, raw_month) = split(v, "-", "a year and a month separated by a dash")?;
+                    let (raw_year, raw_month) =
+                        split(v, "-", "a year and a month separated by a dash")?;
                     fn parse_number<O: FromStr, E: DeError>(raw: &str) -> Result<O, E> {
                         raw.parse::<O>().or_else(|_| {
                             Err(DeError::invalid_value(
@@ -59,14 +66,18 @@ impl<'de> Deserialize<'de> for Period {
                     let year = parse_number(raw_year)?;
                     let month = parse_number(raw_month)?;
 
-                    let start_date = NaiveDate::from_ymd_opt(year, month, 1)
-                        .ok_or(DeError::invalid_value(Unexpected::Str(&v), &"a valid month"))?;
+                    let start_date = NaiveDate::from_ymd_opt(year, month, 1).ok_or(
+                        DeError::invalid_value(Unexpected::Str(&v), &"a valid month"),
+                    )?;
                     let end_date = (start_date + Months::new(1))
                         .pred_opt()
                         .ok_or(DeError::custom("Could not construct month"))?;
                     (start_date, end_date)
                 } else {
-                    return Err(DeError::invalid_length(v.len(), &"a either 7 or 21 characters long string"));
+                    return Err(DeError::invalid_length(
+                        v.len(),
+                        &"a either 7 or 21 characters long string",
+                    ));
                 };
                 Ok(Period {
                     start_date,
@@ -79,26 +90,41 @@ impl<'de> Deserialize<'de> for Period {
 }
 
 impl Serialize for Period {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error> where S: Serializer {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
         if self.start_date.day() == 1 {
             let end_of_month = {
-                let day_after_end = self.end_date.succ_opt().ok_or(SerError::custom("Could not compute day after end of month"))?;
+                let day_after_end = self
+                    .end_date
+                    .succ_opt()
+                    .ok_or(SerError::custom("Could not compute day after end of month"))?;
                 self.end_date.month() != day_after_end.month()
             };
             if end_of_month {
-                return serializer.serialize_str(self.start_date.format("%Y-%m").to_string().as_str());
+                return serializer
+                    .serialize_str(self.start_date.format("%Y-%m").to_string().as_str());
             }
         }
 
-        serializer.serialize_str(format!("{}/{}", self.start_date.format("%F") , self.end_date.format("%F")).as_str())
+        serializer.serialize_str(
+            format!(
+                "{}/{}",
+                self.start_date.format("%F"),
+                self.end_date.format("%F")
+            )
+            .as_str(),
+        )
     }
 }
 
 #[cfg(test)]
 mod test_deserialize {
-    use super::Period;
     use chrono::NaiveDate;
     use serde_test::{assert_tokens, Token};
+
+    use super::Period;
 
     #[test]
     fn test_deserialize_precise_period() {
