+++ src/remaining/mod.rs
@@ -1,0 +1,7 @@
+pub use legacy::{Amount, DisplayAccount, DisplayGoal, Figure, RemainingMoneyScreen};
+pub use operations::RemainingOperation;
+
+mod legacy;
+mod operations;
+mod vault_values;
+
