+++ src/transaction/recurring.rs
@@ -1,9 +1,13 @@
 use std::collections::HashSet;
+
 use serde::Deserialize;
-use crate::period::PeriodsConfiguration;
+
 use transactions::Transaction;
+
 use crate::amounts::{Currency, Figure};
 use crate::period::Period;
+use crate::period::PeriodsConfiguration;
+
 use super::transactions;
 
 #[derive(Deserialize)]
@@ -12,16 +16,16 @@ struct RecurringTransactionsConfiguration<T: PeriodsConfiguration> {
     amount: Figure,
     currency: Currency,
     when: T,
-    paid: HashSet<Period>
+    paid: HashSet<Period>,
 }
 
 impl<T: PeriodsConfiguration> RecurringTransactionsConfiguration<T> {
-    fn generate_predicted_transaction(&self, p: Period) -> Transaction{
-        return Transaction{
+    fn generate_predicted_transaction(&self, p: Period) -> Transaction {
+        return Transaction {
             amount: &self.amount,
             currency: &self.currency,
 
-            occurrences: vec![]
-        }
+            occurrences: vec![],
+        };
     }
 }
