+++ src/transaction/transactions.rs
@@ -1,15 +1,14 @@
-use chrono::NaiveDate;
-use crate::period::Period;
 use crate::amounts::{Currency, Figure};
+use crate::period::Period;
 
 pub struct Transaction<'a> {
     pub amount: &'a Figure,
     pub currency: &'a Currency,
 
-    pub occurrences: Vec<Occurrence<'a>>
+    pub occurrences: Vec<Occurrence<'a>>,
 }
 
-pub struct Occurrence<'a>{
+pub struct Occurrence<'a> {
     period: &'a Period,
-    paid: bool
+    paid: bool,
 }
